Tweaked view so that player is in the centre of the screen
The resolution of the view is 256x192px, i.e 16x12 tiles. As the view
was aligned with the top-left corner of tiles, this meant there were 8
tiles in view to the left of the player and only 7 to the right (similar
vertically). The view has been offset by half a tile in each direction
to correct this.